requirements.txt
google-generativeai>=0.3.2
faster-whisper>=1.0.1
edge-tts>=6.1.9
pyttsx3>=2.90
pyautogui>=0.9.54
pytesseract>=0.3.10
psutil>=5.9.8
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
loguru>=0.7.2
python-dotenv>=1.0.1
wmi>=1.5.1
screen-brightness-control>=0.27.1
pycaw>=20230407
pvporcupine>=3.0.1
sounddevice>=0.4.6
comtypes>=1.3.1
PyAudio>=0.2.14
